refactor: 넥스트 스탭 로또 기계(LottoMachine)에 대한 로직을 구현합니다.
- [X] 문자열을 판독해서 수동 로또 티켓을 만들 수 있다. (구분자: ', ')
- [X] 구매 가능한 만큼 자동 로또 티켓들을 만들 수 있다.

+++ src/main/java/lotto/Money.java
@@ -33,9 +33,9 @@ public class Money {
         return valueOf(amount.add(other.amount));
     }
 
-    public long divide(final Money other) {
+    public int divide(final Money other) {
         return amount.divide(other.amount, RoundingMode.DOWN)
-                .longValue();
+                .intValue();
     }
 
     public double profitRate(final Money other) {
